Increase constness

+++ sado_globals.h
@@ -15,8 +15,7 @@ namespace sado
   extern int outputfreq;
   extern int haploid;
   extern int diploid;
-  //extern bigint popsize;
-  extern bigint endtime;
+  extern int endtime;
   extern double x0;
   extern double p0;
   extern double q0;
